+++ src/vs/workbench/browser/parts/editor/editorParts.ts
@@ -5,26 +5,17 @@
 
 import { localize } from 'vs/nls';
 import { EditorGroupLayout, GroupDirection, GroupLocation, GroupOrientation, GroupsArrangement, GroupsOrder, IAuxiliaryEditorPart, IEditorDropTargetDelegate, IEditorGroupsService, IEditorSideGroup, IFindGroupScope, IMergeGroupOptions } from 'vs/workbench/services/editor/common/editorGroupsService';
-import { Event, Emitter } from 'vs/base/common/event';
+import { Emitter } from 'vs/base/common/event';
 import { getActiveDocument } from 'vs/base/browser/dom';
 import { Disposable, DisposableStore, IDisposable, toDisposable } from 'vs/base/common/lifecycle';
 import { GroupIdentifier } from 'vs/workbench/common/editor';
-import { AuxiliaryEditorPart, EditorPart, MainEditorPart } from 'vs/workbench/browser/parts/editor/editorPart';
+import { EditorPart, MainEditorPart } from 'vs/workbench/browser/parts/editor/editorPart';
 import { IEditorGroupView, IEditorPartsView } from 'vs/workbench/browser/parts/editor/editor';
 import { InstantiationType, registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import { IAuxiliaryWindowOpenOptions, IAuxiliaryWindowService } from 'vs/workbench/services/auxiliaryWindow/browser/auxiliaryWindowService';
-import { ILifecycleService } from 'vs/workbench/services/lifecycle/common/lifecycle';
-import { WindowTitle } from 'vs/workbench/browser/parts/titlebar/windowTitle';
+import { IAuxiliaryWindowOpenOptions } from 'vs/workbench/services/auxiliaryWindow/browser/auxiliaryWindowService';
 import { distinct } from 'vs/base/common/arrays';
-import { IStatusbarService } from 'vs/workbench/services/statusbar/browser/statusbar';
-import { ServiceCollection } from 'vs/platform/instantiation/common/serviceCollection';
-import { EditorStatus } from 'vs/workbench/browser/parts/editor/editorStatus';
-import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { ITitleService } from 'vs/workbench/services/title/browser/titleService';
-import { isNative } from 'vs/base/common/platform';
-import { IAuxiliaryTitlebarPart } from 'vs/workbench/browser/parts/titlebar/titlebarPart';
-import { getTitleBarStyle } from 'vs/platform/window/common/window';
+import { AuxiliaryEditorPart } from 'vs/workbench/browser/parts/editor/auxiliaryEditorPart';
 
 export class EditorParts extends Disposable implements IEditorGroupsService, IEditorPartsView {
 
@@ -33,12 +24,7 @@ export class EditorParts extends Disposable implements IEditorGroupsService, IEd
 	readonly mainPart = this._register(this.createMainEditorPart());
 
 	constructor(
-		@IInstantiationService private readonly instantiationService: IInstantiationService,
-		@IAuxiliaryWindowService private readonly auxiliaryWindowService: IAuxiliaryWindowService,
-		@ILifecycleService private readonly lifecycleService: ILifecycleService,
-		@IConfigurationService private readonly configurationService: IConfigurationService,
-		@IStatusbarService private readonly statusbarService: IStatusbarService,
-		@ITitleService private readonly titleService: ITitleService
+		@IInstantiationService private readonly instantiationService: IInstantiationService
 	) {
 		super();
 
@@ -55,105 +41,15 @@ export class EditorParts extends Disposable implements IEditorGroupsService, IEd
 	readonly onDidCreateAuxiliaryEditorPart = this._onDidCreateAuxiliaryEditorPart.event;
 
 	async createAuxiliaryEditorPart(options?: IAuxiliaryWindowOpenOptions): Promise<IAuxiliaryEditorPart> {
-		const disposables = new DisposableStore();
-
-		// Window
-		const auxiliaryWindow = disposables.add(await this.auxiliaryWindowService.open(options));
-
-		// Status Bar Visibility
-		const statusBarConfiguration = 'workbench.statusBar.visible';
-		let statusBarVisible = this.configurationService.getValue<boolean>(statusBarConfiguration) !== false;
-		disposables.add(this.configurationService.onDidChangeConfiguration(e => {
-			if (e.affectsConfiguration(statusBarConfiguration)) {
-				statusBarVisible = this.configurationService.getValue<boolean>(statusBarConfiguration) !== false;
-
-				updateStatusBarVisibility(true);
-			}
-		}));
-
-		function computeEditorPartHeightOffset(): number {
-			let editorPartHeightOffset = 0;
-
-			if (statusBarVisible) {
-				editorPartHeightOffset += statusbarPart.height;
-			}
-
-			if (titlebarPart) {
-				editorPartHeightOffset += titlebarPart.height;
-			}
-
-			return editorPartHeightOffset;
-		}
-
-		function updateStatusBarVisibility(fromEvent: boolean): void {
-			if (statusBarVisible) {
-				statusbarPart.container.style.display = 'block';
-			} else {
-				statusbarPart.container.style.display = 'none';
-			}
-
-			editorPartContainer.style.height = `calc(100% - ${computeEditorPartHeightOffset()}px)`;
-
-			if (fromEvent) {
-				auxiliaryWindow.layout();
-			}
-		}
-
-		const editorPart = disposables.add(this.instantiationService.createInstance(AuxiliaryEditorPart, auxiliaryWindow.window.vscodeWindowId, this, this.getGroupsLabel(this._parts.size)));
-
-		// Titlebar (part)
-		let titlebarPart: IAuxiliaryTitlebarPart | undefined = undefined;
-		const useCustomTitle = isNative && getTitleBarStyle(this.configurationService) === 'custom'; // custom title in aux windows only enabled in native
-		if (useCustomTitle) {
-			titlebarPart = disposables.add(this.titleService.createAuxiliaryTitlebarPart(auxiliaryWindow.container, editorPart));
-		} else {
-			disposables.add(this.instantiationService.createInstance(WindowTitle, auxiliaryWindow.window, editorPart));
-		}
-
-		// Editor Part
-		const editorPartContainer = document.createElement('div');
-		editorPartContainer.classList.add('part', 'editor');
-		editorPartContainer.setAttribute('role', 'main');
-		auxiliaryWindow.container.appendChild(editorPartContainer);
-
-		disposables.add(this.registerEditorPart(editorPart));
-		editorPart.create(editorPartContainer, { restorePreviousState: false });
-
-		// Statusbar
-		const statusbarPart = disposables.add(this.statusbarService.createAuxiliaryStatusbarPart(auxiliaryWindow.container));
-
-		// Editor status scoped to auxiliary window
-		const scopedInstantiationService = this.instantiationService.createChild(new ServiceCollection(
-			[IStatusbarService, statusbarPart]
-		));
-		disposables.add(scopedInstantiationService.createInstance(EditorStatus, editorPart));
-
-		updateStatusBarVisibility(false);
-
-		// Lifecycle
-		const editorCloseListener = disposables.add(Event.once(editorPart.onWillClose)(() => auxiliaryWindow.window.close()));
-		disposables.add(Event.once(auxiliaryWindow.onWillClose)(() => {
-			if (disposables.isDisposed) {
-				return; // the close happened as part of an earlier dispose call
-			}
-
-			editorCloseListener.dispose();
-			editorPart.close();
-			disposables.dispose();
-		}));
-		disposables.add(Event.once(this.lifecycleService.onDidShutdown)(() => disposables.dispose()));
-
-		// Layout
-		disposables.add(auxiliaryWindow.onDidLayout(dimension => editorPart.layout(dimension.width, dimension.height - computeEditorPartHeightOffset(), 0, 0)));
-		auxiliaryWindow.layout();
+		const { part, disposables } = await this.instantiationService.createInstance(AuxiliaryEditorPart, this).create(this.getGroupsLabel(this._parts.size), options);
 
 		// Events
-		this._onDidAddGroup.fire(editorPart.activeGroup);
+		this._onDidAddGroup.fire(part.activeGroup);
 
 		const eventDisposables = disposables.add(new DisposableStore());
-		this._onDidCreateAuxiliaryEditorPart.fire({ part: editorPart, disposables: eventDisposables });
+		this._onDidCreateAuxiliaryEditorPart.fire({ part, disposables: eventDisposables });
 
-		return editorPart;
+		return part;
 	}
 
 	//#endregion
@@ -163,7 +59,7 @@ export class EditorParts extends Disposable implements IEditorGroupsService, IEd
 	private readonly _parts = new Set<EditorPart>();
 	get parts() { return Array.from(this._parts); }
 
-	private registerEditorPart(part: EditorPart): IDisposable {
+	registerEditorPart(part: EditorPart): IDisposable {
 		this._parts.add(part);
 
 		const disposables = this._register(new DisposableStore());
